[IMP] connector_importer: new option 'use_job' on import type
By default one job is spawned for each data model/importer used in a
import type. It works well if the records to import have no relation
between them, but if we choose to disable this behavior we are able to
link records imported from one importer to another as we are in the same
SQL transaction (e.g. an order and its lines).