[IMP] connector_importer: handle the unique key as an external/XML ID
'importer.record' and 'importer.odoorecord.handler' components have been
updated to handle the 'odoo_unique_key' attribute as an XML-ID if the
new attribute 'odoo_unique_key_is_xmlid' is set to 'True'.
This XML-ID will then be used as usual to find an existing record, and
will be set on the created record if it doesn't already exist.